Nimble MCMC script to fit a random effect Hill model.

Fit a random effect dose response model using Nimble for fast MCMC. Nimble creates a compiled sampler that can run iterations much faster than the manually implemented version.

Usage

run_RE_nimble(
  y_i,
  Cx,
  replicate_sets,
  n_iter = 50000,
  n_burn = 10000,
  n_thin = 20,
  prior_alpha_sd = sqrt(1000),
  prior_theta_shape = 0.001,
  prior_theta_rate = 0.001,
  prior_beta_shape = 0.3,
  prior_beta_rate = 0.3,
  prior_sd_noninfo = 0.1
)

Arguments

y_i

Response data as a matrix where each row is a chemical and each column is a dose.

Cx

Dose data with dimension equal to y_i.

replicate_sets

A list of arrays where each array indicates the rows of y_i that correspond to replicates of a single chemical

n_iter

Number of MCMC iterations to run, defaults to 50k

n_burn

Number of initial MCMC iterations to discard, defaults to 10k or half of n_iter, whichever is lower.

n_thin

Number of MCMC iterations to skip to approximate iid samples, defaults to 20. If it is larger than difference in iterations and burn in, it is reset to 1.

prior_alpha_sd

hyperparameter for the sill, controls prior variance

prior_theta_shape

hyperparameter for the EC50, controls gamma shape

prior_theta_rate

hyperparameter for the EC50, controls gamma rate

prior_beta_shape

hyperparameter for the slope, controls gamma shape

prior_beta_rate

hyperparameter for the slope, controls gamma rate

prior_sd_noninfo

hyperparameter for various noise terms which use a weakly informative prior, controls spread from 1. Default is 0.1.

Value

nimble_samples: a list of MCMC chains with an entry for each parameter of the model.